bug fix by srishailam

+++ partyanalyst-service/trunk/src/main/java/com/itgrids/partyanalyst/service/impl/SurveyDataDetailsService.java
@@ -3965,7 +3965,7 @@ public class SurveyDataDetailsService implements ISurveyDataDetailsService
 			
 			List<Long> surveyCallStatusIds = surveyCallStatusDAO.getSurveyCallDetailsByVoterId(surveyReportVO.getVoterId());
 			
-			Long surveyCallStatusId = (Long) surveyCallStatusIds.get(0);
+			Long surveyCallStatusId = surveyCallStatusIds.get(0);
 			
 			if(surveyCallStatusId != null && surveyCallStatusId != 0 )
 			{
